+++ src/template/multiple.mjs
@@ -1,6 +1,7 @@
 /**
  * {ManagedBanner} banner
  * {Array<BannerData>} data
+ * {Function(String expression, Integer index): String|null} expr
  */
 export default `
     <div class="amp-banner amp-banner--multiple">
@@ -26,7 +27,8 @@ export default `
                                      <% if(null !== banner.positionData.dimensions.width) { %>width="<%- banner.positionData.dimensions.width %>"<% } %>
                                      <% if(null !== banner.positionData.dimensions.height) { %>height="<%- banner.positionData.dimensions.height %>"<% } %>
                                      <% if('' !== b.content.title) { %>title="<%- b.content.title %>"<% } %>
-                                     <% if(banner.options.has('loading') && index >= parseInt(banner.options.get('loading-offset', 0))) { %>loading="<%- banner.options.get('loading') %>"<% } %>>
+                                     <% if(banner.options.has('loading') && index >= parseInt(banner.options.get('loading-offset', 0))) { %>loading="<%- banner.options.get('loading') %>"<% } %>
+                                     <% if(banner.options.has('fetchpriority') && expr(banner.options.get('fetchpriority'), index)) { %>fetchpriority="<%- expr(banner.options.get('fetchpriority'), index) %>"<% } %>>
                             </picture>
                         </a>
                     <% } else if ('html' === b.content.type) { %>