+++ src/template/single.mjs
@@ -1,6 +1,7 @@
 /**
  * {ManagedBanner} banner
  * {BannerData} data
+ * {Function(String expression, Integer index): String|null} expr
  */
 export default `
     <div class="amp-banner amp-banner--single"
@@ -23,7 +24,8 @@ export default `
                          <% if(null !== banner.positionData.dimensions.width) { %>width="<%- banner.positionData.dimensions.width %>"<% } %>
                          <% if(null !== banner.positionData.dimensions.height) { %>height="<%- banner.positionData.dimensions.height %>"<% } %>
                          <% if('' !== data.content.title) { %>title="<%- data.content.title %>"<% } %>
-                         <% if(banner.options.has('loading')) { %>loading="<%- banner.options.get('loading') %>"<% } %>>
+                         <% if(banner.options.has('loading')) { %>loading="<%- banner.options.get('loading') %>"<% } %>
+                         <% if(banner.options.has('fetchpriority') && expr(banner.options.get('fetchpriority'), 0)) { %>fetchpriority="<%- expr(banner.options.get('fetchpriority'), 0) %>"<% } %>>
                 </picture>
             </a>
         <% } else if ('html' === data.content.type) { %>